Formatting +keeping setThreeDLayout as a unique function of ThreeDComponent util file
