+++ src/client/app/components/RadarChartComponent.tsx
@@ -24,7 +24,6 @@ import { lineUnitLabel } from '../utils/graphics';
 import translate from '../utils/translate';
 import SpinnerComponent from './SpinnerComponent';
 import {setHelpLayout} from '../utils/setLayout';
-//import {setThreeDLayout} from '../utils/setLayout';
 
 /**
  * @returns radar plotly component
@@ -253,7 +252,7 @@ export default function RadarChartComponent() {
 						angularaxis: {
 							// TODO Attempts to format the dates to remove the time did not work with plotly
 							// choosing the tick values which is desirable. Also want time if limited time range.
-							direction: 'clockwise',
+							direction: 'counterclockwise',
 							showgrid: true,
 							gridcolor: '#ddd',
 							nticks: maxTicks
